CRUD de Cliente terminado.

+++ src/main/java/com/managent/cafe/Entity/Cliente.java
@@ -9,7 +9,7 @@ import lombok.Data;
 public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idCliente;
+    private Integer idCliente;
 
     private String nombreCliente;
